Move `CoincidentHalfEdgesAreNotSiblings`
Move it to the module for the new validation infrastructure.
