+++ crates/fj-core/src/validation/checks/mod.rs
@@ -2,6 +2,7 @@
 //!
 //! See documentation of [parent module](super) for more information.
 
+mod coincident_half_edges_are_not_siblings;
 mod curve_geometry_mismatch;
 mod face_boundary;
 mod face_winding;
@@ -9,6 +10,7 @@ mod half_edge_connection;
 mod half_edge_has_no_sibling;
 
 pub use self::{
+    coincident_half_edges_are_not_siblings::CoincidentHalfEdgesAreNotSiblings,
     curve_geometry_mismatch::CurveGeometryMismatch,
     face_boundary::FaceHasNoBoundary,
     face_winding::InteriorCycleHasInvalidWinding,
